confirmation page skeleton and code refactoring
- created route to confirmation page without making unnecessary renders
- separated logic to handle reserved dates and times in a separate custom hook called useReserve, left submission of reservation onto backend in useReservationCard
