+++ swimster-ui/src/hooks/useListingDetail.js
@@ -1,15 +1,10 @@
 import { useState, useEffect } from 'react'
 import { useAuthContext, isUserAuthenticated } from "../contexts/auth"
-import { useNavigate } from 'react-router-dom'
-import { useForm } from 'react-hook-form'
 import apiClient from "../services/apiClient"
 
 export const useListingDetail = listingId => {
     const { user, initialized } = useAuthContext()
-    const { register, setValue, handleSubmit } = useForm()
     const [isFetching, setIsFetching] = useState(false)
-    const [isSubmitProcessing, setIsSubmitProcessing] = useState(false)
-    const navigate = useNavigate()
     const [error, setError] = useState(null)
     const [listing, setListing] = useState({})
     const [listingImages, setListingImages] = useState([])
@@ -40,34 +35,10 @@ export const useListingDetail = listingId => {
         }
     }, [listingId, isAuthenticated])
 
-    // handler for making a reservation for the current listing
-    const onSubmit = async formData => {
-        setIsSubmitProcessing(true)
-
-        const formattedFormData = {
-            reservationDate: formData.reservationDate,
-            startTime: formData.reservationStartTime,
-            endTime: formData.reservationEndTime,
-            guests: parseInt(formData.reservationGuests),
-        }
-
-        const { data, error } = await apiClient.createReservation(JSON.stringify(formattedFormData), listingId)
-        if (error) setError(error)
-        if (data?.reservation) {
-            navigate('/')
-        }
-        setIsSubmitProcessing(false)
-    }
-
     return {
         listing,
         listingImages,
         error,
         isFetching,
-        isSubmitProcessing,
-        register,
-        setValue,
-        handleSubmit,
-        onSubmit,
     }
 }
